add: transaction crud, check stock book, refactor remove book and employee

+++ src/entity/Employee.java
@@ -1,14 +1,14 @@
 package entity;
 
 public class Employee {
-    private long nip;
+    private int nip;
     private String name;
 
     public long getNip() {
         return nip;
     }
 
-    public void setNip(long nip) {
+    public void setNip(int nip) {
         this.nip = nip;
     }
